refactor: удалена лишняя мемоизация и вынесены подкомпоненты OfferScreen
